pyproject.toml
[project]
name = "robot-bridge-manager"
version = "0.1.0"
description = "A Python package to generate ROS Gazebo bridge YAML from URDF files."
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer[all]>=0.9.0",
    "lxml>=4.9.0",
    "pyyaml>=6.0",
]

[project.scripts]
gz-bridge-builder = "gz_bridge_builder.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["gz-bridge-builder/src"]

[tool.pytest.ini_options]
pythonpath = ["gz-bridge-builder/src"]

[tool.setuptools.package-data]
"gz_bridge_builder" = ["py.typed"]

[tool.mypy]
ignore_missing_imports = true # Ignore missing imports for now, refine later
plugins = [
    "pydantic.mypy"
]
